Reviewer 2
Overall, this was a very good app. Good Information about yourself. I would advise you to better fill up the blank spaces. Blank spaces do not generally appeal to the user so if you were able to fill up some of the white space, the overall feel of the app would be much better. Resizing a few of the buttons would also be beneficial to make the flow and transitioning a little bit smother. Overall good app.